requirements.txt
fastapi = 0.123.10
uvicorn = 0.38.0
scikit-learn = 1.6.1
pandas = 2.3.3
joblib = 1.5.2
asyncpg = 0.31.0
databases = 0.9.0
psycopg2-binary = 2.9.11
redis = 7.0.1
groq = 0.37.1
python-dotenv = 1.2.1
pydantic-settings = 2.11.0